Add function checkFleetHelmApps
